basic model structure
finished basic model structure and objectId validation.

+++ models/address.model.js
@@ -1,5 +1,9 @@
 const DB = require('../utils/DB');
 
+function isValidObjectId(id) {
+    return id === null || (/^[0-9a-fA-F]{24}$/).test(id);
+}
+
 class AddressModel {
     region;
     city;
@@ -17,7 +21,7 @@ class AddressModel {
         this.apartment = apartment;
         this.notes = notes;
         this.location = {
-            type: "Point",
+            type: "Point", // a data type of mongodb 
             coordinates: [longitude, latitude]
         }
     }
@@ -28,14 +32,23 @@ class AddressModel {
     }
 
     static async read(id) {
+        if (!isValidObjectId(id)) {
+            throw new Error('Invalid ObjectId');
+        }
         return await new DB().findOne('addresses', { _id: id });
     }
 
     static async update(id, updateData) {
+        if (!isValidObjectId(id)) {
+            throw new Error('Invalid ObjectId');
+        }
         return await new DB().updateOne('addresses', { _id: id }, updateData);
     }
 
     static async delete(id) {
+        if (!isValidObjectId(id)) {
+            throw new Error('Invalid ObjectId');
+        }
         return await new DB().deleteOne('addresses', { _id: id });
     }
